--- src/pages/EventInfo/index.js
@@ -1,1 +1,0 @@
-export * from './EventInfo'
