+++ src/pages/EventInfoPage/EventInfoPage.styles.js
