+++ src/pages/EventInfoPage/__test__/EventInfoPage.test.js
