+++ src/pages/SponsorPage/SponsorPage.jsx
@@ -1,11 +1,12 @@
 import React from 'react'
 import { MainTemplate } from '../../templates'
-import { SponsorCard } from '../../components/molecules'
+import { SponsorCard, TitlePage } from '../../components/molecules'
 import { SponsorPageStyled } from './SponsorPage.styles'
 
 export function SponsorPage() {
   return (
-    <MainTemplate title="Sponsors" button="Add sponsor">
+    <MainTemplate>
+      <TitlePage title="Sponsors" button="Add sponsor" />
       <SponsorPageStyled>
         <SponsorCard />
       </SponsorPageStyled>
