+++ src/pages/index.js
@@ -4,4 +4,4 @@ export * from './AgendaPage'
 export * from './SpeakerPage'
 export * from './SponsorPage'
 export * from './BroadcastPage'
-export * from './EventInfo'
+export * from './EventInfoPage'
